+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## 58.2.0
 
 ### Enhancements
+- **API Changes**
+  - Added a survey review page customized for a single subject
 - **Gem Changes**
   - Updated to rails 5.2.0.rc1
   - Updated to pg 1.0.0